Refactor emotes handling
This commit creates an `Emotes` class capable of handling emote-related
operations with a greater level of verification than a simple dict
offers. This is also a step toward decoupling all of the global data
from `dragonbot.py` and making this program a bit more modular (though
the "bot" as an entity dominates this sort of system and makes it
difficult to model in a more modular way).