Add --version option
This option will print the version of the bot (and the underlying
Discord.py API) and exit.

It's possible that even with this option the program still initiates a
connection to the Discord API. Ideally, no network access should be
needed for this option. Testing with my network interface down didn't
cause any problems though.